Finished authentication. Home page redirects to login/sign up page if not logged in.

+++ routes/tasks.js
@@ -1,46 +1,39 @@
 var express = require('express');
 var router = express.Router();
-var Task = require('../models/task');
+var Task = require('../models/task.js');
+
+function isLoggedIn(req, res, next) {
+    if (req.isAuthenticated()){
+        res.locals.username = req.user.local.username;
+        next();
+    }
+    else {
+        res.redirect('/auth')
+    }
+}
+
+router.use(isLoggedIn);
 
 /* GET home page. */
 router.get('/', function(req, res, next) {
-    //Homepage
     //listing all uncompleted tasks
-  Task.tasks.find({completed: false})
+  Task.find({_creator : req.user, completed: false})
       .then((docs) => {
     res.render('index', {title: 'Incomplete tasks', tasks: docs})
       })
       .catch((err) => {next(err)});
 });
 
-/* GET login page. */
-router.get('/login', function (res, req, next) {
-    res.redirect('login');
-});
 
-/* GET signup page. */
-router.get('/signup', function (res, req, next) {
-    res.redirect('signup');
-});
-
-/* POST login page. */
-router.post('/login', passport.authenticate('local-login', {
-    successRedirect: '/secret',
-    failureRedirect: '/login',
-    failureFlash: true
-}));
-
-/* POST signup page. */
-router.post('/signup', passport.authenticate('local-login', {
-    successRedirect: '/secret',
-    failureRedirect: '/signup',
-    failureFlash: true
-}));
 
 router.post('/add', function(req, res, next){
     //creates a new task when the add button is pushed on the homepage.
-  if (req.body.text) {
-      var t = new Task({text: req.body.text, completed: false, dateCreated:new Date()});
+  if (!req.body || !req.body.text){
+      req.flash('error', 'Please enter a task.');
+      res.redirect('/');
+  }
+  else {
+      var t = new Task({ _creator: req.user, text: req.body.text, completed: false, dateCreated:new Date()});
 
       t.save().then((newTask) => {
           console.log('The new task created is ', newTask);
@@ -49,25 +42,19 @@ router.post('/add', function(req, res, next){
           next(err);
       });
   }
-  else{
-      req.flash('error', 'Please enter a task.');
-      res.redirect('/');
-  }
 });
 
 router.post('/done', function(req, res, next){
     //updates individual task completed to true when done button is pushed
 
-  Task.findByIdAndUpdate( req.body._id, {completed: true, dateCompleted:new Date()})
+  Task.findByIdAndUpdate( {_id: req.body._id, _creator: req.user.id}, {completed: true, dateCompleted:new Date()})
       .then( (originalTask) => {
         if(originalTask){
-            req.flash('info', originalTask.text + ' marked as done!');
-          res.redirect('/');
+            res.status(403).send('This is not your task!')
         }
         else{
-          var err= new Error('Not Found');
-          err.status = 404;
-          next(err);
+          req.flash('info', 'Task marked as done');
+          res.redirect('/')
         }
       })
       .catch((err) => {next(err);
@@ -76,25 +63,23 @@ router.post('/done', function(req, res, next){
 
 router.post('/delete', function(req, res, next){
     //deletes an individual task on the homepage
-  Task.findByIdAndRemove(req.body._id)
+  Task.findByIdAndRemove({_id: req.body._id, _creator: req.user.id})
        .then((deletedTask) => {
-     if(deletedTask){
-         req.flash('info','Task deleted');
-       res.redirect('/');
-     }
-     else{
-         var err= new Error('Not Found');
-         err.status = 404;
-         next(err);
-     }
-   })
-       .catch((err) => {next(err);
-       });
+           if(deletedTask){
+               res.status(403).send('This is not your task!')
+           }
+           else{
+               req.flash('info', 'Task deleted');
+               res.redirect('/')
+           }
+       })
+      .catch((err) => {next(err);
+      });
 });
 
 router.post('/alldone', function(req, res, next){
     //updates all tasks on the home page completed key to true.
-    Task.updateMany({completed: false}, {completed: true})
+    Task.updateMany({_creator: req.user, completed: false}, {completed: true}, {multi: true})
         .then(()=>{
         req.flash('info', 'All tasks are done!');
       res.redirect('/')
@@ -105,7 +90,7 @@ router.post('/alldone', function(req, res, next){
 
 router.post('/deleteDone', function(req, res, next){
     //deletes all completed tasks
-    Task.deleteMany({completed: true})
+    Task.deleteMany({_creator: req.user, completed: true})
         .then(() => {
             req.flash('info', 'All completed tasks deleted');
             res.redirect('/')
@@ -116,7 +101,7 @@ router.post('/deleteDone', function(req, res, next){
 
 router.get('/completed', function(req, res, next){
     //finds and renders all completed tasks onto the completed_task view
-  Task.find({completed:true})
+  Task.find({creator: req.user._id, completed:true})
       .then((docs) => {
     res.render('completed_tasks', {tasks: docs });
       }).catch((err) => {next(err);
@@ -127,36 +112,21 @@ router.get('/task/:_id', function(req, res, next){
     //creates an individual task page when clicked
   Task.findById(req.params._id)
       .then((doc) => {
-    if(doc){
-      res.render('task', {task: doc});
+    if(!doc){
+        res.status(404).send('Task not found.');
+    }
+    else if(!doc._creator.equals(req.user._id)){
+        res.status(403).send('This is not your task!');
     }
     else{
-      next();
+        res.render('task', {task: doc});
     }
+
       })
       .catch((err) => {
     next(err);
       });
 });
 
-/* GET secret page*/
-router.get('/secret', isLoggedIn, function (req,res,next){
-
-    var user = req.task.local;
-
-    res.render('secret', {
-        username : user.username
-    });
-
-});
-
-function isLoggedIn(req, res, next){
-    if (req.isAuthenticated()){
-        next();
-    }
-    else{
-        res.redirect('/login');
-    }
-}
 
 module.exports = router;
